Some more size_t...

+++ common/phone/nokia/nfunc.h
@@ -88,8 +88,8 @@ GSM_Error N71_65_DecodePhonebook		(GSM_StateMachine	*s,
   				 		 unsigned char 		*MessageBuffer,
   				 		 int 			MessageLength,
 						 bool			DayMonthReverse);
-int 		N71_65_EncodePhonebookFrame	(GSM_StateMachine *s, unsigned char *req, GSM_MemoryEntry *entry, int *block2, bool DCT4, bool VoiceTag);
-int 		N71_65_PackPBKBlock		(GSM_StateMachine *s, int id, int size, int no, unsigned char *buf, unsigned char *block);
+size_t 		N71_65_EncodePhonebookFrame	(GSM_StateMachine *s, unsigned char *req, GSM_MemoryEntry *entry, size_t *block2, bool DCT4, bool VoiceTag);
+size_t 		N71_65_PackPBKBlock		(GSM_StateMachine *s, int id, size_t size, int no, unsigned char *buf, unsigned char *block);
 
 #endif
